Revert "[JIMT][CT-446] upload files to codebridge (#61403)"
This reverts commit cfc8059820e7d7e2a83764e2145dcdf6470211e2.
