+++ apps/i18n/codebridge/en_us.json
@@ -1,4 +1,5 @@
 {
+  "addFile": "Add file",
   "addSubFolder": "Add sub-folder",
   "areYouSure": "Are you sure?",
   "allTestsPassed": "All tests passed!",
@@ -10,29 +11,27 @@
   "clearConsole": "Clear console",
   "closeFile": "close file",
   "defaultLayout": "Default layout",
-  "defaultNewFileContents": "Add your changes to {fileName}",
   "delete": "Delete",
-  "deleteFile": "Delete",
+  "deleteFile": "Delete file",
   "deleteFileConfirm": "Are you sure you want to delete the file \"{filename}\"?",
-  "deleteFolder": "Delete",
+  "deleteFolder": "Delete folder",
   "deleteFolderConfirm": "Are you sure you want to delete the folder \"{folderName}\"?",
   "deleteFolderConfirmBoth": "This will also delete {fileCount} files and {folderCount} subfolders inside \"{folderName}\".",
   "deleteFolderConfirmFiles": "This will also delete {fileCount} files inside \"{folderName}\".",
   "deleteFolderConfirmSubfolders": "This will also delete {folderCount} subfolders inside \"{folderName}\".",
-  "downloadFile": "Download",
+  "downloadFile": "Download file",
   "downloadFileError": "Cannot download files of type {extension}",
   "duplicateFileError": "Filename {fileName} is already in use in this folder. Please choose a different name.",
   "duplicateFolderError": "Foldername {folderName} is already in use in this folder. Please choose a different name.",
   "duplicateSupportFileError": "Filename {fileName} is already in use in this folder in the level's support code. Please choose a different name.",
   "folderExistsError": "Folder already exists with that name",
-  "invalidFileUpload": "Cannot upload files of type {fileType}",
-  "invalidNameError": "Names must be only letters, numbers, or underscores",
+  "invalidNameError": "Names must be only letters, numbers, or dots",
   "loadingEnvironmentTooltip": "Your environment is loading. Please wait.",
   "makeStarter": "Make starter file",
   "makeSupport": "Make support file",
   "makeValidation": "Make validation file",
-  "moveFile": "Move",
-  "moveFolder": "Move",
+  "moveFile": "Move file",
+  "moveFolder": "Move folder",
   "moveFilePrompt": "Please enter your destination folder",
   "moveFolderPrompt": "Please enter your destination folder",
   "newFile": "New File",
@@ -42,8 +41,8 @@
   "noFileExtensionError": "Please include a file extension",
   "noOpenFiles": "Open a file to get started!",
   "predictRunDisabledTooltip": "You must answer the question in the instructions before running the code.",
-  "renameFile": "Rename",
-  "renameFolder": "Rename",
+  "renameFile": "Rename file",
+  "renameFolder": "Rename folder",
   "restoredOldVersion": "Version successfully restored",
   "result": "Result",
   "rootFolder": "project root",
@@ -55,7 +54,6 @@
   "testName": "Test Name",
   "testsDidNotPass": "Tests did not all pass.",
   "toggleFileBrowser": "Toggle File Browser",
-  "uploadFile": "Upload",
   "validate": "Validate",
   "validatingRunDisabledTooltip": "You cannot run code while validation is running",
   "validationNotYetPassedContinue": "You must pass tests before continuing.",
