+++ apps/src/codebridge/FileBrowser/FileBrowserHeaderPopUpButton.tsx
@@ -1,5 +1,3 @@
-import {useCodebridgeContext} from '@codebridge/codebridgeContext';
-import {DEFAULT_FOLDER_ID} from '@codebridge/constants';
 import {PopUpButton} from '@codebridge/PopUpButton/PopUpButton';
 import classNames from 'classnames';
 import React from 'react';
@@ -7,8 +5,6 @@ import React from 'react';
 import codebridgeI18n from '@cdo/apps/codebridge/locale';
 import FontAwesomeV6Icon from '@cdo/apps/componentLibrary/fontAwesomeV6Icon/FontAwesomeV6Icon';
 
-import {FileUploader} from './FileUploader';
-import {useFileUploadErrorCallback, useHandleFileUpload} from './hooks';
 import {newFolderPromptType, newFilePromptType} from './types';
 
 import moduleStyles from './styles/filebrowser.module.scss';
@@ -22,56 +18,27 @@ type FileBrowserHeaderPopUpButtonProps = {
 export const FileBrowserHeaderPopUpButton = ({
   newFolderPrompt,
   newFilePrompt,
-}: FileBrowserHeaderPopUpButtonProps) => {
-  const {
-    project,
-    config: {validMimeTypes},
-  } = useCodebridgeContext();
-  const uploadErrorCallback = useFileUploadErrorCallback();
-  const handleFileUpload = useHandleFileUpload(project.files);
-  return (
-    <PopUpButton iconName="plus" alignment="left">
-      <div
-        onClick={() => newFolderPrompt()}
-        className={classNames(
-          darkModeStyles.dropdownItem,
-          moduleStyles.dropdownItem
-        )}
-      >
-        <FontAwesomeV6Icon iconName="plus" iconStyle="solid" />
-        <div>{codebridgeI18n.newFolder()}</div>
-      </div>
-      <div
-        onClick={() => newFilePrompt()}
-        className={classNames(
-          darkModeStyles.dropdownItem,
-          moduleStyles.dropdownItem
-        )}
-      >
-        <FontAwesomeV6Icon iconName="plus" iconStyle="solid" />
-        <div>{codebridgeI18n.newFile()}</div>
-      </div>
-      <FileUploader
-        validMimeTypes={validMimeTypes}
-        callback={(fileName, contents) =>
-          handleFileUpload({
-            folderId: DEFAULT_FOLDER_ID,
-            fileName,
-            contents,
-          })
-        }
-        errorCallback={uploadErrorCallback}
-      >
-        <div
-          className={classNames(
-            darkModeStyles.dropdownItem,
-            moduleStyles.dropdownItem
-          )}
-        >
-          <FontAwesomeV6Icon iconName="upload" iconStyle="solid" />
-          <div>{codebridgeI18n.uploadFile()}</div>
-        </div>
-      </FileUploader>
-    </PopUpButton>
-  );
-};
+}: FileBrowserHeaderPopUpButtonProps) => (
+  <PopUpButton iconName="plus" alignment="left">
+    <div
+      onClick={() => newFolderPrompt()}
+      className={classNames(
+        darkModeStyles.dropdownItem,
+        moduleStyles.dropdownItem
+      )}
+    >
+      <FontAwesomeV6Icon iconName="plus" iconStyle="solid" />
+      <div>{codebridgeI18n.newFolder()}</div>
+    </div>
+    <div
+      onClick={() => newFilePrompt()}
+      className={classNames(
+        darkModeStyles.dropdownItem,
+        moduleStyles.dropdownItem
+      )}
+    >
+      <FontAwesomeV6Icon iconName="plus" iconStyle="solid" />
+      <div>{codebridgeI18n.newFile()}</div>
+    </div>
+  </PopUpButton>
+);
