--- apps/src/codebridge/FileBrowser/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useFileUploadErrorCallback';
-export * from './useHandleFileUpload';
