+++ apps/src/codebridge/types.ts
@@ -51,7 +51,6 @@ export type ConfigType = {
   };
   activeGridLayout?: string;
   showFileBrowser: boolean;
-  validMimeTypes?: string[];
 };
 
 export type ProjectType = MultiFileSource;
