--- apps/src/codebridge/utils/validateFileName.ts
@@ -1,3 +1,0 @@
-// restrict typed in input to what we consider to be valid names, which for now are word characters or hyphens + a dot + more word characters
-export const validateFileName = (name: string = '') =>
-  Boolean(name.match(/^[\w-]+\.\w+$/));
