--- apps/src/codebridge/utils/viewableImageFileType.ts
@@ -1,8 +1,0 @@
-const defaultViewableImageFileTypesArray = ['png', 'jpg', 'gif'];
-export const viewableImageFileType = (
-  extension: string,
-  viewableImageFileTypesArray = defaultViewableImageFileTypesArray
-) => {
-  const viewableImageFileTypes = new Set(viewableImageFileTypesArray);
-  return viewableImageFileTypes.has(extension);
-};
