+++ apps/src/metrics/AnalyticsConstants.js
@@ -414,12 +414,6 @@ const EVENTS = {
   CODEBRIDGE_RENAME_FILE: 'Rename file on codebridge',
   CODEBRIDGE_RENAME_FOLDER: 'Rename folder on codebridge',
   CODEBRIDGE_DOWNLOAD_FILE: 'Download file on codebridge',
-  CODEBRIDGE_UPLOAD_FILE: 'Upload file on codebridge',
-  CODEBRIDGE_UPLOAD_INVALID_FILE_NAME:
-    'Attempted upload of invalid file name on codebridge',
-  CODEBRIDGE_UPLOAD_UNACCEPTED_FILE:
-    'Attempted upload of unaccepted file on codebridge',
-  CODEBRIDGE_UPLOAD_FAILED: 'Failed to upload file on codebridge',
 
   // Codebridge - Other events
   CODEBRIDGE_CLEAR_CONSOLE: 'Console cleared on codebridge',
