+++ apps/src/pythonlab/PythonlabView.tsx
@@ -94,7 +94,6 @@ const defaultConfig: ConfigType = {
   labeledGridLayouts,
   activeGridLayout: 'horizontal',
   showFileBrowser: true,
-  validMimeTypes: ['text/'],
 };
 
 const PythonlabView: React.FunctionComponent = () => {
